If one int is INT64, so must everything else be.
Thanks to @tclune for help with this. Was fun. Every timer was negative!

+++ base/MAPL_Profiler.F90
@@ -119,7 +119,7 @@
 
       character(len=ESMF_MAXSTR), parameter :: IAm="MAPL_ProfClockOff"
 
-      integer :: COUNTS
+      integer(kind=INT64) :: COUNTS
       integer :: I, NN
       integer :: status
